class of things that can be split
At the core is the Split object, which carries functions for splitting the input into two pieces at some given split factor, how to recombine the pieces, and the dimension along which the split took place.

Instances return the ways in which the input can be split. For examples, scalars can not be split, vectors (and lists) can be split in only one way, and matrices can be split either horizontally or vertically.